requirements.txt
djangorestframework
django-rest-knox